fix: rename project name
modified:   README.md
modified:   setup.py
new file:   src/qjam/__init__.py
renamed:    src/vaqm/q_array.py -> src/qjam/q_array.py
renamed:    src/vaqm/q_clean.py -> src/qjam/q_clean.py
renamed:    src/vaqm/q_sub.py -> src/qjam/q_sub.py
renamed:    src/vaqm/q_util.py -> src/qjam/q_util.py
renamed:    src/vaqm/vaqm.py -> src/qjam/qjam.py
