+++ README.md
@@ -1,2 +1,2 @@
-# vaqm
+# qjam
 UGE qsub wrapper for using qsub from commandline.
