+++ src/qjam/q_sub.py
@@ -7,7 +7,7 @@ __author__ = "Riku Okajima<riku.okajima@gmail.com>"
 __date__ = "2018-06-04"
 import argparse
 from typing import List
-from vaqm.q_util import (
+from qjam.q_util import (
     aquire_script_path,
     create_file,
     escape_string,
